feat: hook useAquariumControls
